Сделал экран информации о займе, небольшой рефакторинг

+++ app/src/main/java/com/khomichenko/sergey/homework1410/presentation/main_loan_screen/fragment/recycler_view/MainLoanAdapter.kt
@@ -7,7 +7,7 @@ import com.khomichenko.sergey.homework1410.databinding.LoanItemBinding
 import com.khomichenko.sergey.homework1410.domain.entity.main_loan.LoanEntity
 
 class MainLoanAdapter(
-    private val onViewClickListener:() -> Unit
+    private val onViewClickListener:(Int) -> Unit
 ) : RecyclerView.Adapter<MainLoanViewHolder>() {
 
     var loanList = emptyList<LoanEntity>()
